give deprecation its own integration test
