+++ test/integration/001_simple_copy_test/models/advanced_incremental.sql
@@ -9,7 +9,7 @@
 select *
 from {{ ref('seed') }}
 
-{% if adapter.already_exists(this.schema, this.table) and not flags.FULL_REFRESH %}
+{% if is_incremental() %}
 
     where id > (select max(id) from {{this}})
