+++ test/integration/012_deprecation_tests/models/sql_where.sql
@@ -1,0 +1,3 @@
+{{ config(sql_where='id > (select max(id) from {{this}})')}}
+
+select 1 as id
